Move token limit logics from controller to concern
